+++ src/i18n/index.ts
@@ -1,5 +1,10 @@
-import { i18n } from './i18n';
+import { hasLocaleMessage, i18n } from './i18n';
 import { preactTranslator } from './preactTranslator';
 import { translator } from './translator';
 
-export { translator, preactTranslator, i18n };
+export {
+    translator,
+    preactTranslator,
+    i18n,
+    hasLocaleMessage,
+};