+++ src/pages/common/components/Icons/index.tsx
@@ -1,0 +1,1 @@
+export { ArrowIcon, CloseIcon, ShieldIcon } from './Icons';